Install phlex gem for object oriented views
I haven't used it before, but I'm liking the idea of pure ruby for
designing partials and views. I was considering Slim or (my usual)
Haml, but decided to give this a try.

Note: The generator installed a phlex application layout as well,
but I don't think I want to use it so I removed it.